added relation check between user and role when deleting role

+++ src/main/java/com/hossein/PermissionTree/service/impl/role/RoleService.java
@@ -8,17 +8,22 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.hossein.PermissionTree.controller.viewModel.Role.RoleViewModel;
+import com.hossein.PermissionTree.controller.viewModel.User.UserViewModel;
 import com.hossein.PermissionTree.dao.repository.role.IRoleRepository;
 import com.hossein.PermissionTree.dto.role.RoleDto;
 import com.hossein.PermissionTree.exception.ApplicationException;
 import com.hossein.PermissionTree.model.role.Role;
 import com.hossein.PermissionTree.service.role.IRoleService;
+import com.hossein.PermissionTree.service.user.IUserService;
 
 @Service
 public class RoleService implements IRoleService {
 	
 	@Autowired
 	private IRoleRepository iRoleRepository;
+	
+	@Autowired
+	private IUserService iUserService;
 
 	@Override
 	public List<Role> getAllRoles() {
@@ -44,6 +49,9 @@ public class RoleService implements IRoleService {
 	@Override
 	@Transactional
 	public Boolean delete(Long id) {
+		List<UserViewModel> users = this.iUserService.getUsersByRoleId(id);
+		if (users.size() > 0) 
+			throw new ApplicationException("Relation Exists Between Role And User");
 		this.iRoleRepository.deleteById(id);
 		return this.load(id) == null;
 	}
